+++ src/utils/CollectionUtils.js
@@ -1,6 +1,6 @@
 // Node.js imports
 import { dirname, join } from 'node:path'
-import parse from 'parse-gitignore'
+import gitignore from 'parse-gitignore'
 
 // Archie Imports
 import Components from '../config/Components.js'
@@ -29,13 +29,10 @@ class CollectionUtils {
    * Get Watch Folders for a Collection
    * @param collection
    */
-  static async getIgnorePatterns (collection) {
-    let gitIgnoreContents
-    let gitIgnorePatterns
+  static getIgnorePatterns (collection) {
     const ignorePatterns = IGNORE_PATTERNS
     if (collection.gitIgnoreFile) {
-      gitIgnoreContents = await FileUtils.getFileContents(join(collection.gitIgnoreFile))
-      gitIgnorePatterns = parse(gitIgnoreContents)
+      const gitIgnorePatterns = gitignore.parse(collection.gitIgnoreFile).patterns
       ignorePatterns.push(...gitIgnorePatterns)
     }
     return ignorePatterns