+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.7.2] - 2023-11-01
+
+### Fixed
+
+- Fixed the computation of gitignore patterns during watch mode ([#228](https://github.com/archetype-themes/archie/pull/228))
+
+## [1.7.1] - 2023-10-31
+
+### Changed
+
+- Refactor Import map processor and add comments
+
 ## [1.7.0] - 2023-10-24
 
 ### Added